user schemas url bug, added authentication to delete quiz route.

+++ app/api/quiz/[id]/route.ts
@@ -1,5 +1,5 @@
 import { WithAuth } from '@/lib/api/auth-protected';
-import { handleError } from '@/lib/api/errors';
+import { APIError, handleError } from '@/lib/api/errors';
 import { prisma } from '@/lib/prisma';
 import { NextResponse } from 'next/server';
 
@@ -24,14 +24,24 @@ export const GET = WithAuth(async (req, { user, params }) => {
 /*
  * Delete a quiz based on its ID.
  */
-export const DELETE = WithAuth(async (req, { params }) => {
+export const DELETE = WithAuth(async (req, { user, params }) => {
   try {
     const { id } = await params;
-    await prisma.quiz.delete({
-      where: {
-        id,
-      },
+    const quiz = await prisma.quiz.findFirst({
+      where: { id },
     });
+
+    if (!quiz) {
+      throw new APIError('Invalid Quiz ID', 404);
+    } else {
+      if (quiz.createdBy == user.id) {
+        await prisma.quiz.delete({
+          where: { id },
+        });
+      } else {
+        throw new APIError('User must be owner of the quiz.', 401);
+      }
+    }
 
     return NextResponse.json({ sucess: true }, { status: 200 });
   } catch (error) {
